fix(runtime): address prompt payload review feedback
